Move components from Root component to App
