+++ src/containers/Root.js
@@ -1,19 +1,7 @@
 import React, { Component } from 'react';
 import { Provider } from 'react-redux';
-import AddDoto from './AddDoto';
-import DotoList from './DotoList';
-import ClockControl from './ClockControl';
+import App from './App';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import Paper from 'material-ui/Paper';
-
-const paperStyle = {
-  width: 600,
-  padding: 20,
-  marginTop: 50,
-  marginBottom: 50,
-  marginLeft: 'auto',
-  marginRight: 'auto'
-};
 
 export default class Root extends Component {
   render() {
@@ -21,11 +9,7 @@ export default class Root extends Component {
     return (
       <MuiThemeProvider>
         <Provider store={ store }>
-          <Paper style={ paperStyle } zDepth={ 1 }>
-            <ClockControl />
-            <AddDoto />
-            <DotoList />
-          </Paper>
+          <App />
         </Provider>
       </MuiThemeProvider>
     );
